refactor: 결제 타입 데이터 제외

+++ src/main/java/roomescape/utility/PaymentClient.java
@@ -4,5 +4,5 @@ import roomescape.dto.business.PaymentResult;
 
 public interface PaymentClient {
 
-    PaymentResult pay(String paymentKey, String orderId, int amount);
+    PaymentResult pay(String paymentKey, String orderId, long amount);
 }
